refactored for more meaningful distribution of repsonsibilities across files
