+++ src/cli/options.go
@@ -1,27 +1,12 @@
 package main
 
 import (
-  "log"
   "fmt"
   "github.com/jroimartin/gocui"
-  "gopkg.in/yaml.v2"
-  "strconv"
-  "io/ioutil"
   "time"
   "strings"
+  "strconv"
 )
-
-// MODULE ARGS
-
-type Arg struct {
-  name string
-  required bool
-  input string
-}
-
-func (a Arg) String() string {
-  return "name: " + a.name + " input: " + a.input + " required: " + strconv.FormatBool(a.required)
-}
 
 // OPTION INTERFACE
 
@@ -158,107 +143,4 @@ func (to saveOption) Name() string {
 
 func (to saveOption) IsModuleConfig() bool {
   return false
-}
-
-// MTRIAGE INTERFACE
-
-func getNextOption(g *gocui.Gui, newState state) option {
-
-  if newState.folder == "" {
-    return textInputOption{ prompt: "please enter the path to your working directory", name: "folder", validationType: TYPE_FOLDER  }
-  }
-
-  if newState.phase == "" {
-    return multiOption{ options: []string{"select","analyse"}, name: "phase" }
-  }
-
-  if newState.currentModule == "" {
-    modules := modulesForPhase(newState.phase)
-    return multiOption{ options: modules, name: "module" }
-  }
-
-  configOptions := configOptionsForModule(newState.currentModule, newState.phase)
-  moduleConfig := newState.configs[newState.currentModule].(map[string]interface{})
-  existingKeys := keysForMap(moduleConfig)
-  for i := range configOptions {
-    option := configOptions[i]
-    exists := false
-    for j := range existingKeys {
-      key := existingKeys[j]
-      if option.name == key {
-        exists = true
-      }
-    }
-
-    if !exists {
-      name := option.name
-      input := option.input
-      prompt := "please enter a " + input + " for argument:\n\n   "  + name
-      return textInputOption { prompt: prompt, name: name, validationType: input }
-    }
-  }
-  return saveOption{ isComposable: newState.phase == PHASE_ANALYSE }
-}
-
-func modulesForPhase(phase string) []string {
-  path, _ := phaseToLibFolder(phase)
-  return dirNamesIn(path, []string{"meta"})
-}
-
-func phaseToLibFolder(phase string) (path string, name string) {
-  if phase == PHASE_SELECT {
-    return DIR_SELECTORS, LIB_SELECTORS
-  } else if phase == PHASE_ANALYSE {
-    return DIR_ANALYSERS, LIB_ANALYSERS
-  } else {
-    log.Panicln("invalid phase!")
-    return "", ""
-  }
-}
-
-func configOptionsForModule(module string, phase string) []Arg {
-
-  argsPath := argsPathForModule(module, phase)
-
-  file, err := ioutil.ReadFile(argsPath)
-  if err != nil {
-    log.Panicln(err)
-  }
-
-  var argsWild []map[string]interface{}
-  err = yaml.Unmarshal(file, &argsWild)
-  if err != nil {
-    log.Panicln(err)
-  }
-
-  argsTame := []Arg{}
-
-  if phase == PHASE_ANALYSE {
-    els_in := Arg{ name: "elements_in", input: TYPE_WHITELIST, required: true }
-    argsTame = append(argsTame, els_in)
-  }
-
-  for i := range argsWild {
-    name, ok := argsWild[i]["name"].(string)
-    if !ok {
-      log.Panicln("invalid args yaml! name not string")
-    }
-    input, ok := argsWild[i]["input"].(string)
-    if !ok {
-      log.Panicln("invalid args yaml! input not string")
-    }
-    required, ok := argsWild[i]["required"].(bool)
-    if !ok {
-      log.Panicln("invalid args yaml! required not bool")
-    }
-    arg := Arg{ name: name, input: input, required: required }
-    argsTame = append(argsTame, arg)
-  }
-
-  return argsTame
-}
-
-func argsPathForModule(module string, phase string) string {
-  phaseDir, _ := phaseToLibFolder(phase)
-  return phaseDir + "/" + module + "/args.yaml"
 }
